many to many

+++ RECAP/articles/views.py
@@ -9,9 +9,13 @@ from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 def index(request):
+    visits_num = request.session.get('visits', 0)
+    request.session['visits'] = visits_num + 1
+    request.session.modified = True
     articles = Article.objects.all()
     context = {
         'articles' : articles,
+        'visits' : visits_num,
     }
     return render(request, 'articles/index.html', context)
 
@@ -55,13 +59,14 @@ def create(request):
     #     return redirect('accounts:login')
     if request.method == 'POST':
         form = ArticleForm(request.POST)
-        embed()
         # 전송된 데이터가 유효한 값인지 검사
         if form.is_valid():
             # title = form.cleaned_data.get('title')
             # content = form.cleaned_data.get('content')
             # article = Article.objects.create(title=title,content=content)
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
         # 사실 위 처럼 적을 필요 없이
         # form.save()
         # 하면 자동으로 저장됨
@@ -102,22 +107,26 @@ def create(request):
 @login_required
 def update(request, pk):
     article = get_object_or_404(Article, pk=pk)
-    if request.method == 'POST':
-        # 실제 DB의 데이터를 수정
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            # article.title = form.cleaned_data.get('title')
-            # article.content = form.cleaned_data.get('content')
-            # article.save()
-            form.save()
-            return redirect(article)
-        else:
-            form = ArticleForm(
-            initial={
-                'title': article.title,
-                'content': article.content,
-            })
-            return redirect(article)
+
+    if article.user == request.user:
+        if request.method == 'POST':
+            # 실제 DB의 데이터를 수정
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                # article.title = form.cleaned_data.get('title')
+                # article.content = form.cleaned_data.get('content')
+                # article.save()
+                form.save()
+                return redirect(article)
+            else:
+                form = ArticleForm(
+                initial={
+                    'title': article.title,
+                    'content': article.content,
+                })
+                return redirect(article)
+    else:
+        return redirect('articles:index')
     # 편집화면
     
     # form = ArticleForm(
@@ -141,9 +150,12 @@ def delete(request, pk):
     if request.user.is_authenticated:
         # POST로 요청이 올 때만, 지우는 걸로
         article = get_object_or_404(Article, pk=pk)
-        if request.method == 'POST':
-            article.delete()
-            return redirect('articles:index')
+        if article.user == request.user:
+            if request.method == 'POST':
+                article.delete()
+                return redirect('articles:index')
+            else:
+                return redirect(article)
         else:
             return redirect(article)
     # 401 에러를 리턴
@@ -156,6 +168,7 @@ def create_comment(request, pk):
         comment_form = CommentForm(request.POST)
         if comment_form.is_valid():
             comment = comment_form.save(commit=False) # commit은 데이터베이스 커밋. 커밋을 안 하고 저장부터 하겠다. 그러면 저장 직전에 객체를 만든다.
+            comment.user = request.user
             comment.article_id = pk # DB에는 실제로 article_id 로 저장되어 있음
             comment.save() # 이렇게 하면 실제 DB에 반영이 된다.
             ###################
@@ -163,4 +176,25 @@ def create_comment(request, pk):
             # comment.article = article
 
     # return redirect('articles:detail', article_pk)
+    return redirect(article)
+
+
+def send_cookie(request):
+    # 항상 우리는 return render를 했는데 이는 Httpresponse를 return 하는 것
+    # return redirect() (내부적으로 reverse를 사용)
+    # return reverse()
+    # return HttpResponse()
+    res = HttpResponse('과자 받아라')
+    res.set_cookie('mycookie', 'oreo')
+    return res
+
+def like(request, article_pk):
+    # article_pk로 넘어온 글의 like_users에 현재 접속 중인 유저를 추가
+    # request.user.like_articles.add(Article.objects.get(pk=article_pk))
+    article = Article.objects.get(pk=article_pk)
+    if request.user in article.like_users.all():
+        article.like_users.remove(request.user)
+    else:
+        # request.user.like_articles.add(article)
+        article.like_users.add(request.user)
     return redirect(article)
